Trait snowbridge_outbound_queue_runtime_api::OutboundQueueApi

pub trait OutboundQueueApi<Block: BlockT, Balance>: Core<Block>where
    Balance: BalanceT,{
    // Provided methods
    fn prove_message(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        leaf_index: u64
    ) -> Result<Option<MerkleProof>, ApiError> { ... }
    fn calculate_fee(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        message: Message
    ) -> Result<Option<Balance>, ApiError> { ... }
}

Provided Methods

fn prove_message( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, leaf_index: u64 ) -> Result<Option<MerkleProof>, ApiError>

Generate a merkle proof for a committed message identified by leaf_index. The merkle root is stored in the block header as a \[sp_runtime::generic::DigestItem::Other\]

fn calculate_fee( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, message: Message ) -> Result<Option<Balance>, ApiError>

Calculate the delivery fee for message